Starting to build 3d version for github. Adding dataloding using dataframes not json. Saving augmented data as test. Added model and 3d language cross attention

+++ segmentation_3d/dataloader_3d.py
@@ -91,9 +91,13 @@ class TextImageDataset(Dataset):
         img_path = self.data.image[index]
         #print(img_name)
 
-        pet_img = self.load_nii_to_numpy(img_path).astype(np.float32)
-        ct_img = self.load_nii_to_numpy(self.data.image2[index]).astype(np.float32)
-        label = self.load_nii_to_numpy(self.data.label[index]).astype(np.float32)
+        #pet_img = self.load_nii_to_numpy(img_path).astype(np.float32)
+        #ct_img = self.load_nii_to_numpy(self.data.image2[index]).astype(np.float32)
+        #label = self.load_nii_to_numpy(self.data.label[index]).astype(np.float32)
+
+        pet_img = self.load_nii_to_numpy(img_path)
+        ct_img = self.load_nii_to_numpy(self.data.image2[index])
+        label = self.load_nii_to_numpy(self.data.label[index])
 
         pet_img = np.expand_dims(pet_img, axis=0)
         ct_img  = np.expand_dims(ct_img, axis=0)
